06.04.2023
İstatistik Pdf'i yaratılıyor.

+++ core/internal/domain/dto/sell_statistics_dto.go
@@ -4,7 +4,7 @@ import "time"
 
 type SellStaticsDto struct {
 	OperationNumber string
-	Balance         float64
+	Balance         float32
 	OrderId         int32
 	ProductName     string
 	OrderQuantity   int32
