+++ core/internal/domain/dto/wallet_dto.go
@@ -1,5 +1,5 @@
 package dto
 
 type WalletDto struct {
-	Balance float64 `json:"balance" validate:"gte=1,number"`
+	Balance float32 `json:"balance" validate:"gte=1,number"`
 }
